# Conflicts:
#	src/main/java/com/fc/controller/DormitoryManagerController.java
#	src/main/java/com/fc/service/DormitoryManagerService.java
#	src/main/java/com/fc/service/impl/DormitoryManagerServiceImpl.java

+++ src/main/java/com/fc/service/impl/DormitoryServiceImpl.java
@@ -1,13 +1,11 @@
 package com.fc.service.impl;
 
-import com.fc.dao.BuildingMapper;
 import com.fc.dao.DormitoryMapper;
-import com.fc.entity.Building;
-import com.fc.entity.Dormitory;
-import com.fc.entity.DormitoryExample;
-import com.fc.entity.DormitoryManager;
+import com.fc.dao.LiveMapper;
+import com.fc.entity.*;
 import com.fc.service.DormitoryService;
 import com.fc.util.RandomNum;
+import com.fc.vo.DataVO;
 import com.fc.vo.DormitoryVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,6 +17,9 @@ import java.util.List;
 public class DormitoryServiceImpl implements DormitoryService {
     @Autowired
     private DormitoryMapper dormitoryMapper;
+
+    @Autowired
+    private LiveMapper liveMapper;
 
     @Override
     public List findAll() {
@@ -45,8 +46,26 @@ public class DormitoryServiceImpl implements DormitoryService {
 
 
     @Override
-    public void delete(String id) {
-        dormitoryMapper.deleteByPrimaryKey(id);
+    public DataVO delete(String id) {
+        LiveExample liveExample = new LiveExample();
+
+        LiveExample.Criteria criteria = liveExample.createCriteria();
+
+        criteria.andDormitoryIdEqualTo(id);
+
+        List<Live> lives = liveMapper.selectByExample(liveExample);
+
+        DataVO dataVO = null;
+
+        if (lives.size() == 0 || lives == null) {
+            dormitoryMapper.deleteByPrimaryKey(id);
+
+            dataVO = new DataVO(200, "删除宿舍成功", true);
+        } else {
+            dataVO = new DataVO(400, "宿舍有人，无法删除", false);
+        }
+
+        return dataVO;
     }
 
     @Override
